// frontend/packages/@domains/geolocation/src/api/proto/ws_message.ts
export enum Channel {
  GPS_REALTIME = 0,
  UNRECOGNIZED = -1,
}
// ...
export interface Request {
  channel: Channel;
  requestId: string;
  publish?: PublishMessage | undefined;
  subscribe?: SubscribeMessage | undefined;
  unsubscribe?: UnsubscribeMessage | undefined;
}

export interface PublishMessage {
  data: Uint8Array;
}

export interface SubscribeMessage {
  indexes: number[];
}

export interface UnsubscribeMessage {
  indexes: number[];
}

export interface Response {
  response?: ResponseMessage | undefined;
  broadcast?: BroadcastMessage | undefined;
}

export interface ResponseMessage {
  requestId: string;
  statusCode: number;
  errorMessage: string;
  data: Uint8Array;
}

export interface BroadcastMessage {
  channel: Channel;
  data: Uint8Array;
}

export interface MessageBatch {
  data: Uint8Array[];
}

function createBaseRequest(): Request {
  return { channel: 0, requestId: "", publish: undefined, subscribe: undefined, unsubscribe: undefined };
}

export const Request: MessageFns<Request> = {
  encode(message: Request, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.channel !== 0) {
      writer.uint32(8).int32(message.channel);
    }
    if (message.requestId !== "") {
      writer.uint32(18).string(message.requestId);
    }
    if (message.publish !== undefined) {
      PublishMessage.encode(message.publish, writer.uint32(26).fork()).join();
    }
    if (message.subscribe !== undefined) {
      SubscribeMessage.encode(message.subscribe, writer.uint32(34).fork()).join();
    }
    if (message.unsubscribe !== undefined) {
      UnsubscribeMessage.encode(message.unsubscribe, writer.uint32(42).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Request {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.channel = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.requestId = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.publish = PublishMessage.decode(reader, reader.uint32());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.subscribe = SubscribeMessage.decode(reader, reader.uint32());
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.unsubscribe = UnsubscribeMessage.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Request {
    return {
      channel: isSet(object.channel) ? channelFromJSON(object.channel) : 0,
      requestId: isSet(object.requestId)
        ? globalThis.String(object.requestId)
        : isSet(object.request_id)
        ? globalThis.String(object.request_id)
        : "",
      publish: isSet(object.publish) ? PublishMessage.fromJSON(object.publish) : undefined,
      subscribe: isSet(object.subscribe) ? SubscribeMessage.fromJSON(object.subscribe) : undefined,
      unsubscribe: isSet(object.unsubscribe) ? UnsubscribeMessage.fromJSON(object.unsubscribe) : undefined,
    };
  },

  toJSON(message: Request): unknown {
    const obj: any = {};
    if (message.channel !== 0) {
      obj.channel = channelToJSON(message.channel);
    }
    if (message.requestId !== "") {
      obj.requestId = message.requestId;
    }
    if (message.publish !== undefined) {
      obj.publish = PublishMessage.toJSON(message.publish);
    }
    if (message.subscribe !== undefined) {
      obj.subscribe = SubscribeMessage.toJSON(message.subscribe);
    }
    if (message.unsubscribe !== undefined) {
      obj.unsubscribe = UnsubscribeMessage.toJSON(message.unsubscribe);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Request>, I>>(base?: I): Request {
    return Request.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Request>, I>>(object: I): Request {
    const message = createBaseRequest();
    message.channel = object.channel ?? 0;
    message.requestId = object.requestId ?? "";
    message.publish = (object.publish !== undefined && object.publish !== null)
      ? PublishMessage.fromPartial(object.publish)
      : undefined;
    message.subscribe = (object.subscribe !== undefined && object.subscribe !== null)
      ? SubscribeMessage.fromPartial(object.subscribe)
      : undefined;
    message.unsubscribe = (object.unsubscribe !== undefined && object.unsubscribe !== null)
      ? UnsubscribeMessage.fromPartial(object.unsubscribe)
      : undefined;
    return message;
  },
};

function createBasePublishMessage(): PublishMessage {
  return { data: new Uint8Array(0) };
}

export const PublishMessage: MessageFns<PublishMessage> = {
  encode(message: PublishMessage, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.data.length !== 0) {
      writer.uint32(10).bytes(message.data);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): PublishMessage {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePublishMessage();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.data = reader.bytes();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PublishMessage {
    return { data: isSet(object.data) ? bytesFromBase64(object.data) : new Uint8Array(0) };
  },

  toJSON(message: PublishMessage): unknown {
    const obj: any = {};
    if (message.data.length !== 0) {
      obj.data = base64FromBytes(message.data);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<PublishMessage>, I>>(base?: I): PublishMessage {
    return PublishMessage.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<PublishMessage>, I>>(object: I): PublishMessage {
    const message = createBasePublishMessage();
    message.data = object.data ?? new Uint8Array(0);
    return message;
  },
};

function createBaseSubscribeMessage(): SubscribeMessage {
  return { indexes: [] };
}

export const SubscribeMessage: MessageFns<SubscribeMessage> = {
  encode(message: SubscribeMessage, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    writer.uint32(10).fork();
    for (const v of message.indexes) {
      writer.int64(v);
    }
    writer.join();
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SubscribeMessage {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSubscribeMessage();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag === 8) {
            message.indexes.push(longToNumber(reader.int64()));

            continue;
          }

          if (tag === 10) {
            const end2 = reader.uint32() + reader.pos;
            while (reader.pos < end2) {
              message.indexes.push(longToNumber(reader.int64()));
            }

            continue;
          }

          break;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): SubscribeMessage {
    return {
      indexes: globalThis.Array.isArray(object?.indexes) ? object.indexes.map((e: any) => globalThis.Number(e)) : [],
    };
  },

  toJSON(message: SubscribeMessage): unknown {
    const obj: any = {};
    if (message.indexes?.length) {
      obj.indexes = message.indexes.map((e) => Math.round(e));
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<SubscribeMessage>, I>>(base?: I): SubscribeMessage {
    return SubscribeMessage.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SubscribeMessage>, I>>(object: I): SubscribeMessage {
    const message = createBaseSubscribeMessage();
    message.indexes = object.indexes?.map((e) => e) || [];
    return message;
  },
};
// ...
export const Response: MessageFns<Response> = {
  encode(message: Response, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.response !== undefined) {
      ResponseMessage.encode(message.response, writer.uint32(10).fork()).join();
    }
    if (message.broadcast !== undefined) {
      BroadcastMessage.encode(message.broadcast, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Response {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.response = ResponseMessage.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.broadcast = BroadcastMessage.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Response {
    return {
      response: isSet(object.response) ? ResponseMessage.fromJSON(object.response) : undefined,
      broadcast: isSet(object.broadcast) ? BroadcastMessage.fromJSON(object.broadcast) : undefined,
    };
  },

  toJSON(message: Response): unknown {
    const obj: any = {};
    if (message.response !== undefined) {
      obj.response = ResponseMessage.toJSON(message.response);
    }
    if (message.broadcast !== undefined) {
      obj.broadcast = BroadcastMessage.toJSON(message.broadcast);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Response>, I>>(base?: I): Response {
    return Response.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Response>, I>>(object: I): Response {
    const message = createBaseResponse();
    message.response = (object.response !== undefined && object.response !== null)
      ? ResponseMessage.fromPartial(object.response)
      : undefined;
    message.broadcast = (object.broadcast !== undefined && object.broadcast !== null)
      ? BroadcastMessage.fromPartial(object.broadcast)
      : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}